# gateway/openapi.yaml
openapi: "3.1.0"
info:
  title: Sentinel Gateway API
  version: "1.0.0"
  description: |
    Real-time AI agent governance layer. Intercepts tool calls, evaluates them
    against policies, and returns allow/block/review decisions with a full audit trail.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://sentinel.internal
    description: Production

# =============================================================================
# Security
# =============================================================================
security:
  - AgentApiKey: []
components:
  securitySchemes:
    AgentApiKey:
      type: apiKey
      in: header
      name: X-Sentinel-Agent-Key
      description: >
        API key issued at agent registration. Format: `snl_<hex>`.
        Exchange for a short-lived JWT via POST /v1/auth/token.
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Shared schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Verdict:
      type: string
      enum: [ALLOWED, BLOCKED, HUMAN_REVIEW]
    DecisionPath:
      type: string
      enum: [fast_path, cognitive_path]

    # --- Core Request/Response -----------------------------------------------
    ToolCallRequest:
      type: object
      required: [tool_name, arguments]
      properties:
        tool_name:
          type: string
          minLength: 1
          maxLength: 128
          example: execute_payment
        arguments:
          type: object
          description: Tool-specific arguments. Validated against policy schemas.
          example:
            amount: 75000
            currency: USD
            recipient: vendor_abc
        session_id:
          type: string
          description: Optional. Groups related tool calls in the audit graph.
          example: sess_abc123
        context:
          $ref: "#/components/schemas/RequestContext"
        metadata:
          $ref: "#/components/schemas/RequestMetadata"
    RequestContext:
      type: object
      description: Optional context used by the cognitive path judge.
      properties:
        task_description:
          type: string
          maxLength: 1000
          example: "Process Q1 vendor invoice as per contract clause 4.2"
        conversation_history:
          type: array
          maxItems: 20
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant, system]
              content:
                type: string
                maxLength: 2000
        source_documents:
          type: array
          maxItems: 5
          items:
            type: object
            properties:
              name:
                type: string
              excerpt:
                type: string
                maxLength: 2000
    RequestMetadata:
      type: object
      properties:
        agent_version:
          type: string
          example: "1.2.0"
        request_id:
          type: string
          format: uuid
          description: Client-generated idempotency key.
        timestamp:
          type: string
          format: date-time
    DecisionResponse:
      type: object
      required: [decision_id, verdict, path, latency_ms, policy_version]
      properties:
        decision_id:
          type: string
          example: dec_01hx2k3m4n5p
        verdict:
          $ref: "#/components/schemas/Verdict"
        reason:
          type: string
          description: Human-readable explanation of the decision.
          example: "Blocked: recipient matches denylist pattern admin@"
        path:
          $ref: "#/components/schemas/DecisionPath"
        latency_ms:
          type: number
          format: float
          example: 7.3
        policy_version:
          type: string
          description: Semver of the active policy that evaluated this request.
          example: "data-access-v3.1.0"
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Judge confidence (cognitive path only).
        rate_limit:
          $ref: "#/components/schemas/RateLimitInfo"
    PendingDecisionResponse:
      type: object
      description: Returned with HTTP 202 when cognitive path exceeds sync budget.
      required: [decision_id, status, poll_url, estimated_ms]
      properties:
        decision_id:
          type: string
          example: dec_01hx2k3m4n5p
        status:
          type: string
          enum: [pending]
        poll_url:
          type: string
          example: /v1/decisions/dec_01hx2k3m4n5p
        estimated_ms:
          type: integer
          example: 5000
    RateLimitInfo:
      type: object
      properties:
        tokens_remaining:
          type: integer
          example: 45
        tokens_max:
          type: integer
          example: 50
        reset_at:
          type: string
          format: date-time

    # --- Agent Registration --------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: rate_limit_exceeded
        message:
          type: string
          example: Token bucket exhausted. Retry after reset_at.
        details:
          type: object

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /v1/tool-calls:
    post:
      operationId: evaluateToolCall
      summary: Evaluate a tool call against active policies
      description: |
        The primary Sentinel endpoint. Agents send their intended tool calls here
        before executing them. Sentinel returns ALLOWED / BLOCKED / HUMAN_REVIEW.

        **Sync response (HTTP 200):** Decision made within the fast-path or cognitive
        path time budget.

        **Async response (HTTP 202):** Cognitive path exceeded the sync budget.
        Poll `poll_url` until the decision is ready.
      tags: [Tool Calls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCallRequest"
            examples:
              payment_high_risk:
                summary: High-risk payment (triggers cognitive path)
                value:
                  tool_name: execute_payment
                  arguments:
                    amount: 75000
                    currency: USD
                    recipient: vendor_abc
                  session_id: sess_demo_001
                  context:
                    task_description: "Pay Q1 invoice per contract clause 4.2"
                    source_documents:
                      - name: contract_vendor_abc.pdf
                        excerpt: "Clause 4.2: Payment of $75,000 USD due Q1..."
              file_read_low_risk:
                summary: Low-risk file read (fast path)
                value:
                  tool_name: read_file
                  arguments:
                    path: /data/public/catalog.json
      responses:
        "200":
          description: Decision made synchronously.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResponse"
        "202":
          description: Decision pending (cognitive path async). Poll poll_url.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingDecisionResponse"
        "401":
          description: Missing or invalid API key / JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Agent rate limit exceeded.
          headers:
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request body failed schema validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/decisions/{decision_id}:
    get:
      operationId: getDecision
      summary: Poll for an async decision result
      tags: [Tool Calls]
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Decision is ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResponse"
        "202":
          description: Still pending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingDecisionResponse"
        "404":
          description: Decision ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Agent Registry
  # ---------------------------------------------------------------------------
...